requirements.txt
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.5
cryptography==45.0.5
customtkinter==5.2.2
darkdetect==0.8.0
Django==5.2.4
django-celery-beat==2.8.1
django-extensions==4.1
django-filter==25.1
django-timezone-field==7.1
django-widget-tweaks==1.5.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
msoffcrypto-tool==5.4.2
olefile==0.47
packaging==25.0
parameterized==0.9.0
pillow==11.3.0
promise==2.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-nmap==0.7.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scapy==2.6.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.7.0